feat(client): Strict validation of email and uid on force_auth
I always felt keeping a list of errors that go to the 400 page broke
some sort of abstraction. Instead of keeping a list, a new field
is added to the error config in the error tables, `errorPageBaseUrl`.
Each error also has a `getErrorPageUrl` function.

If an error does not have an `errorPageBaseUrl`, the error will cause
a redirect to the 500 page. If an error does contain an `errorPageBaseUrl`,
it will be used as the base URL and any necessary metadata will be added
to the URL.

fixes #3040